fix tests

+++ GaelO2/app/Jobs/QcReport/SeriesReport.php
@@ -52,7 +52,7 @@ class SeriesReport
     public function deletePreviewImages()
     {
         foreach ($this->previewImagePath as $imagePath) {
-            unlink($imagePath);
+            if ($imagePath != null) unlink($imagePath);
         }
     }
